+++ kamehouse-shell/bin/common/sudoers/www-data/su.sh
@@ -10,6 +10,10 @@ EXIT_PROCESS_CANCELLED=4
 
 main() {
   COMMAND="/home/${KAMEHOUSE_USER}/programs/kamehouse-shell/bin/common/sudoers/www-data/exec-script.sh $@"
+  
+  if [[ "$@" =~ ^"kamehouse/get-kamehouse-secret.sh -s ".* ]]; then
+    COMMAND="/home/${KAMEHOUSE_USER}/programs/kamehouse-shell/bin/$@"
+  fi
 
   /usr/bin/su - ${KAMEHOUSE_USER} -c "${COMMAND}"
 }